v0.2.4 - Added some minor fixed for multi platform compability - Updated CMakeLists.txt library lookup properties for linux - Fixed naming issues with in functions - Fixed problem when parsking path that contains both Windows and Unix slashes - Fixed threading class - Fixed issue with JSON parser when empty JSON array generates an empty node at some special cases - ffw::Monitor fixed - Added method for easier setting of texture wrapping - Added more tests for media, data, and math module - Fixed some warnings reported by OS X Clang - Imporved compilation speed by removing windows.h - Further improved compilation speed by removing some redurant include(s)

+++ source/media/objloader.cpp
@@ -4,6 +4,9 @@
 #define VECTOR_BUCKET_SIZE 256
 #include <string.h>
 #include <limits.h>
+#include "ffw/math/functions.h"
+#include "ffw/math/vec2.h"
+#include "ffw/math/stringmath.h"
 
 ///=============================================================================
 ffw::ObjLoader::ObjLoader(){
@@ -117,7 +120,7 @@ inline void insetValue(std::vector<ffw::Vec3f>& vec, ffw::Vec3f& val) {
 }
 
 inline float fastFloat(const char* begin) {
-	return atof(begin);
+	return (float)atof(begin);
 }
 
 inline int fastAtoi(const char* begin) {
